v0.1.0.0 Initial manifest.json file & default icon
The manifest.json file has the following: a title, a description of what the extension will do, and all of the icons I will be incorporating into the extension as well as their file "default" names. ---> (I haven't actually downloaded the pictures yet).

The default icon will be the extension icon that the user will click to receive a window with a redirecting button I will add later on.